ultrabot/requirements.txt
vk_api
aiosqlite
python-dotenv
aiohttp
